Convert : Fixed resize for 16bits texture. Fixed texture not 4x4 multiple with BCn compression. Fix BC1 alpha with texConv.

+++ hxd/fs/Convert.hx
@@ -477,11 +477,7 @@ class CompressIMG extends Convert {
 		"RGBA" => "R8G8B8A8_UNORM",
 		"R16U" => "R16_UNORM",
 		"RG16U" => "R16G16_UNORM",
-		"RGBA16U" => "R16G16B16A16_UNORM"
-	];
-
-	// Use texconv for BCn format if source is 16bits
-	static var TEXCONV_FMT_16BITS = [
+		"RGBA16U" => "R16G16B16A16_UNORM",
 		"BC1" => "BC1_UNORM",
 		"BC3" => "BC3_UNORM",
 		"BC4" => "BC4_UNORM",
@@ -494,21 +490,11 @@ class CompressIMG extends Convert {
 	}
 
 	function runTexconv(srcPath:String, dstPath:String, args:Array<String>) {
-		// texconv can only handle output dir, and it prepends to srcPath
-		var dstExt = new haxe.io.Path(dstPath).ext;
-		var tmpPath = new haxe.io.Path(dstPath);
-		tmpPath.ext = "tmp." + new haxe.io.Path(srcPath).ext;
-		var tmpFile = tmpPath.toString();
-		try sys.FileSystem.deleteFile(tmpFile) catch (e:Dynamic) {};
-		try sys.FileSystem.deleteFile(dstPath) catch (e:Dynamic) {};
-		sys.io.File.copy(srcPath, tmpFile);
-		command("texconv", ["-y", "-nologo"].concat(args).concat([tmpFile]));
-		tmpPath.ext = "tmp." + dstExt;
-		var p = tmpPath.toString();
-		if ( sys.FileSystem.exists(p) )
-			sys.FileSystem.rename(p, dstPath);
-		// delete after rename in case the format is the same
-		try sys.FileSystem.deleteFile(tmpFile) catch (e:Dynamic) {};
+		if ( srcPath != dstPath ) {
+			try sys.FileSystem.deleteFile(dstPath) catch (e:Dynamic) {};
+			sys.io.File.copy(srcPath, dstPath);
+		}
+		command("texconv", ["-y", "-nologo"].concat(args).concat([dstPath]));
 		return dstPath;
 	}
 
@@ -522,44 +508,58 @@ class CompressIMG extends Convert {
 		}
 		function tempFile(path: String, tag: String, ext=null) : String {
 			var spath = new haxe.io.Path(path);
-			var tmp = '${spath.file}.$tag.' + (ext != null ? ext : spath.ext);
+			var tmp = Sys.getEnv("TEMP") + '${spath.file}.$tag.' + (ext != null ? ext : spath.ext);
 			if(cleanupFiles == null) cleanupFiles = [];
 			cleanupFiles.push(tmp);
 			return tmp;
 		}
 
+		var cachedImage : hxd.res.Image = null;
+		inline function getSrc() {
+			if (cachedImage == null)
+				cachedImage = makeImage(srcPath);
+			return cachedImage;
+		}
+		inline function isSrc16Bits() {
+			return getSrc().getPixelFormat().match(R16F|R16U|RG16F|RG16U|RGB16F|RGB16U|RGBA16F|RGBA16U);
+		}
+		inline function isSrc4x4aligned() {
+			var info = getSrc().getInfo();
+			return (info.width & 3) == 0 && (info.height & 3) == 0;
+		}
+
+		var dstFmt : String = getParam("format");
 		var mips = hasParam("mips") && getParam("mips") == true;
 		if (hasParam("size")) {
 			var maxSize = getParam("size");
-			var image = makeImage(srcPath);
+			var image = getSrc();
 			var pxls = image.getPixels();
+			var srcFmt = Std.string(image.getPixelFormat());
+			var fmt = TEXCONV_FMT.get(srcFmt) ?? "RGBA";
 			if (pxls.width == pxls.height && pxls.width > maxSize) {
 				var resized = tempFile(srcPath, "resized", "tga");
 				var ssize = "" + maxSize;
-				var args = ["-m", "0", "-sepalpha", "-w", ssize, "-h", ssize, "-f", "RGBA", "-ft", "tga"];
-				if ( hasParam("filter") ) {
-					args.push("-if");
-					args.push(getParam("filter"));
-				}
+				var args = ["-m", "0", "-sepalpha", "-w", ssize, "-h", ssize, "-f", fmt, "-ft", "tga"];
+				if (hasParam("filter"))
+					args = args.concat(["-if", getParam("filter")]);
 				runTexconv(srcPath, resized, args);
 				srcPath = resized;
+				cachedImage = null;
 			}
 		}
-		var format = getParam("format");
-		var tcFmt = TEXCONV_FMT.get(format);
-		if (tcFmt == null) {
-			tcFmt = TEXCONV_FMT_16BITS.get(format);
-			if (tcFmt != null) {
-				var image = makeImage(srcPath);
-				if (!image.getInfo().pixelFormat.match(R16F|R16U|RG16F|RG16U|RGB16F|RGB16U|RGBA16F|RGBA16U))
-					tcFmt = null;
-			}
-		}
+		var isBCFormat = dstFmt.indexOf("BC") == 0;
+		var fix4x4 = isBCFormat && !isSrc4x4aligned();
+		var useTc = !isBCFormat || fix4x4 || isSrc16Bits();
+		var tcFmt = useTc ? TEXCONV_FMT.get(dstFmt) : null;
 		if (tcFmt != null) {
 			var args = ["-f", tcFmt, "-srgb"];
 			if (!mips)
 				args = ["-m", "1"].concat(args);
+			if (hasParam("alpha") && dstFmt == "BC1")
+				args = ["-at", ""+ Std.parseInt("" + getParam("alpha")) / 256.0].concat(args);
 			runTexconv(srcPath, dstPath, args);
+			if (fix4x4)
+				runTexconv(dstPath, dstPath, ["-fixbc4x4"]);
 			return cleanup();
 		}
 		var path = new haxe.io.Path(srcPath);
@@ -589,7 +589,7 @@ class CompressIMG extends Convert {
 				srcPath = oldPath;
 				var convertPixels = [];
 				for (layer in 0...info.layerCount) {
-					var layerPath = dstPath + path.file + "_" + layer + "_dds_" + format + "." + path.ext;
+					var layerPath = dstPath + path.file + "_" + layer + "_dds_" + dstFmt + "." + path.ext;
 					var image = makeImage(layerPath);
 					for (mip in 0...info.mipLevels) {
 						var pixels = image.getPixels(null, mip);
@@ -600,7 +600,7 @@ class CompressIMG extends Convert {
 				var convertBytes = hxd.Pixels.toDDSLayers(convertPixels);
 				for (pixels in convertPixels)
 					pixels.dispose();
-				var tmpPath = dstPath + path.file + "_" + format + "." + path.ext;
+				var tmpPath = dstPath + path.file + "_" + dstFmt + "." + path.ext;
 				sys.io.File.saveBytes(tmpPath, convertBytes);
 				return cleanup();
 			}
@@ -617,7 +617,7 @@ class CompressIMG extends Convert {
 			tmpPath = tempFile(dstPath, "tmp", "dds");
 			sys.io.File.saveBytes(tmpPath, sys.io.File.getBytes(srcPath));
 		}
-		if (hasParam("alpha") && format == "BC1")
+		if (hasParam("alpha") && dstFmt == "BC1")
 			args = args.concat(["-DXT1UseAlpha", "1", "-AlphaThreshold", "" + getParam("alpha")]);
 		args = args.concat(["-fd", "" + getParam("format"), tmpPath == null ? srcPath : tmpPath, dstPath]);
 		command("CompressonatorCLI", args);
